CI: Add update workflow
Adding a GitHub workflow to run flapak-external-data-checker every 30 minutes, which should get us similar behaviour
as the current checks run by flathubbot.

This only needs to be in the default branch.
This covers both master and beta branches, so if the beta branch is kept alive, and a new beta app is not submitted,
then we probably want this in.
Collaborators can trigger the workflow manually, though it's not really needed with the current cron schedule.
Auto-merging open PRs is implemented in f-e-d-c's code, so it should work with this workflow and the currently
configured `automerge-flathubbot-prs` property.

This is similar to the example workflow in https://github.com/flathub/flatpak-external-data-checker, but from
testing this in a private repo, I figured that we don't need the somewhat random number prefix to github-actions'
email.

Signed-off-by: rany <ranygh@riseup.net>